+++ client/src/api/PixiManager.ts
@@ -3,7 +3,7 @@ import * as PIXI from 'pixi.js';
 import { PixiObject } from '../app/Pixi/PixiObject';
 import { FONT, loadTextures } from '../app/Pixi/TextureLoader';
 import { ResourceBars } from '../app/Pixi/ResourceBars';
-import { Shop } from '../app/Pixi/Shop';
+import { Shop } from '../app/Pixi/Shop/Shop';
 import { GameBoard } from '../app/Pixi/GameBoard/GameBoard';
 import { MouseManager } from '../app/Pixi/MouseManager';
 import AbstractGameManager from './AbstractGameManager';
