+++ client/src/app/Pixi/MouseManager.ts
@@ -1,7 +1,7 @@
 import { PixiManager } from '../../api/PixiManager';
 import { PieceType } from '../../_types/global/GlobalTypes';
 import { PieceObject } from './PieceObject';
-import { BoardCoords, MoveAttack } from './PixiTypes';
+import { BoardCoords, MoveAttack } from './@PixiTypes';
 import { compareBoardCoords, idxsIncludes } from './PixiUtils';
 import { Ship } from './Ship';
 import { Submarine } from './Submarine';
