+++ client/src/app/Pixi/ShipSprite.ts
@@ -2,7 +2,7 @@ import { PixiManager } from '../../api/PixiManager';
 import { PixiObject } from './PixiObject';
 import * as PIXI from 'pixi.js';
 import { PieceType } from '../../_types/global/GlobalTypes';
-import { PlayerColor } from './PixiTypes';
+import { PlayerColor } from './@PixiTypes';
 import { playerShader } from './Shaders';
 import { SHIPS, SPRITE_W } from './TextureLoader';
