+++ client/src/app/Pixi/Submarine.ts
@@ -2,7 +2,7 @@ import * as PIXI from 'pixi.js';
 import { PixiManager } from '../../api/PixiManager';
 import { ZKPiece, isKnown } from '../../_types/global/GlobalTypes';
 import { PieceObject } from './PieceObject';
-import { CanvasCoords, PlayerColor } from './PixiTypes';
+import { CanvasCoords, PlayerColor } from './@PixiTypes';
 import { boardCoordsFromLoc } from './PixiUtils';
 
 export const SUB_X = 6;
